requirements.txt
asgiref==3.5.2
Django==3.2.7
django-braces==1.14.0
django-cleanup==6.0.0
django-crispy-forms==1.9.1
django-filter==22.1
fontawesomefree==6.1.1
MouseInfo==0.1.3
mss==6.1.0
mysql-connector-python==8.0.29
mysqlclient==2.1.0
numpy==1.21.4
opencv-python==4.5.4.60
pandas==1.4.2
protobuf==4.21.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
python-dateutil==2.8.2
python-dotenv==0.20.0
pytweening==1.0.4
pytz==2021.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.1
xlrd==2.0.1
XlsxWriter==3.1.2
